21.08.24 게시물 페이징 작업. 이전 클릭시 빈칸나올 때까지, 다음 클릭시 끝을 넘어가 에러 발생

+++ src/main/java/com/memo/post/bo/PostBO.java
@@ -1,6 +1,7 @@
 package com.memo.post.bo;
 
 import java.io.IOException;
+import java.util.Collections;
 import java.util.List;
 
 import org.slf4j.Logger;
@@ -16,6 +17,8 @@ import com.memo.post.model.Post;
 @Service
 public class PostBO {
 	
+	
+
 	private Logger log = LoggerFactory.getLogger(this.getClass());
 	
 	@Autowired
@@ -24,8 +27,40 @@ public class PostBO {
 	@Autowired
 	private FileManagerService fileManagerService;
 	
-	public List<Post> getPostListByUserId(int userId){
-		return postDAO.selectPostListByUserId(userId);
+	private static final int POST_MAX_SIZE = 3;
+	
+	public List<Post> getPostListByUserId(int userId, Integer prevId, Integer nextId){
+		// 게시글 번호	10 9 8 | 7 6 5 | 4 3 2 | 1
+		// 1) 다음 : 가장 작은 수(오른쪽 값) => netxIdParam 쿼리 : nextIdParam보다 작은 3개(Limit)을 가져온다.
+		// 2) 이전 : 가장 큰 수(왼쪽 값) => privIdParam 쿼리 : preIdParam 보다 큰 3개(Limit)를 가져온다. 순서가 뒤집히므로 코드에서 정렬을 뒤집는다.
+		
+		String direction = null;
+		Integer standardId = null;
+		if(prevId != null) {
+			// 이전 버튼 클릭
+			direction = "prev";
+			standardId = prevId;
+			
+			List<Post> postList = postDAO.selectPostListByUserId(userId, direction, standardId, POST_MAX_SIZE);
+			Collections.reverse(postList);
+			return postList;
+		}else if(nextId != null) {
+			direction = "next";
+			standardId = nextId;
+		}
+		
+		return postDAO.selectPostListByUserId(userId, direction, standardId, POST_MAX_SIZE);
+	}
+	
+	// 가장 오른쪽 페이지인가?
+	public boolean isLastPage(int userId, int nextId) {
+		//1
+		return nextId == postDAO.selectPostIdByUserIdAndSort(userId, "ASC");
+	}
+	
+	public boolean isFirstPage(int userId, int prevId) {
+		//10
+		return prevId == postDAO.selectPostIdByUserIdAndSort(userId, "DESC");
 	}
 	
 	public Post getPostByPostIdAndUserId(int postId, int userId) {
@@ -74,10 +109,12 @@ public class PostBO {
 		Post post = postDAO.selectPostByPostIdAndUserId(postId, userId);
 		String imagePath = null;
 		imagePath = post.getImagePath(); // url가져오기
-		try {
-			fileManagerService.deleteFile(imagePath);;
-		} catch (IOException e) {
-			log.error("[파일삭제] 삭제 중 에러 : "+postId+" "+ imagePath);
+		if(imagePath!=null) {
+			try {
+				fileManagerService.deleteFile(imagePath);;
+			} catch (IOException e) {
+				log.error("[파일삭제] 삭제 중 에러 : "+postId+" "+ imagePath);
+			}
 		}
 		postDAO.postDelete(userId, postId);
 	}
